enum ActorResult and enum FailurePhase has been moved to `actor_result.rs`.
